Make endpoint keys the route itself instead of its purpose
This is also necessary for the new 429 handling, where different methods don't
matter for the same route (GET /guilds/{id}/channels will be the same as
POST /guilds/{id}/channels).

The naming conventions for each of the routes is simply the path of the route
(ignoring query parameters) with slashes replaced with underscores. ID
parameters are replaced with {first_letter_of_object}id, i. e. a route that
would take a :guild_id parameter has "gid" in that place. The @me in some routes
becomes simply "me"; other parameters (currently only invite "code"s) are left
untouched.

An example: the hypothetical route