Support millisecond precision rate limiting
This changes out HTTP requests to send X-RateLimit-Precision as
"millisecond". With this, Discord will send us millisecond accurate time
values for rate limit headers. Additionally, we will now use
X-RateLimit-Reset-After instead of the HTTP date header, which is only
precise to seconds.

If X-RateLimit-Reset-After is not present, we fall back to Retry-After,
otherwise we panic.

Closes #192