Clarify & fix semantics of Client#reconnect
This patch renames the argument of Client#reconnect to should_resume.
The concept of a "suspended" session is too "meta" - this should get the
point across.

This argument was also broken! Even if you set should_suspend, the
on_close handler would suspend the session anyways. This fixes it by
invalidating the session. It may be more "accurate" to unset @session,
but I'm not sure. For now, we'll keep the data in memory..